repo reorg: rename hcm -> http_connection_manager
My abreviation was not popular.
