+++ source/extensions/filters/network/http_connection_manager/config.cc
@@ -1,4 +1,4 @@
-#include "extensions/filters/network/hcm/config.h"
+#include "extensions/filters/network/http_connection_manager/config.h"
 
 #include <chrono>
 #include <memory>
